Refactor: Improve project structure and modularity.
- Standardized Python package structure with code in `src/chorderizer`.
- Removed duplicate `chorderizer.py` from the root directory.
- Refactored the main `chorderizer.py` script into logical modules:
    - `theory_utils.py`: Core music theory classes.
    - `generators.py`: Chord, tablature, and MIDI generation classes.
    - `ui.py`: User interface components and functions.
- Updated imports across the refactored modules.
- Consolidated dependency management by removing redundant `requeriments.txt` (relying on `setup.py`).
- Updated `README.md` to reflect new structure, correct run commands, and suggest future enhancements like unit testing.
